remove pyarrow requirement from write_database

+++ py-polars/polars/dataframe/frame.py
@@ -4246,14 +4246,12 @@ class DataFrame:
             msg = f"write_database `if_table_exists` must be one of {{{allowed}}}, got {if_table_exists!r}"
             raise ValueError(msg)
 
+        connection_module_root = type(connection).__module__.split(".", 1)[0]
+
         if engine is None:
-            if (
-                isinstance(connection, str)
-                or (module_root := type(connection).__module__.split(".", 1)[0])
-                == "sqlalchemy"
-            ):
+            if isinstance(connection, str) or connection_module_root == "sqlalchemy":
                 engine = "sqlalchemy"
-            elif module_root.startswith("adbc"):
+            elif connection_module_root.startswith("adbc"):
                 engine = "adbc"
 
         def unpack_table_name(name: str) -> tuple[str | None, str | None, str]:
@@ -4268,20 +4266,32 @@ class DataFrame:
             return catalog, schema, tbl  # type: ignore[return-value]
 
         if engine == "adbc":
-            adbc_driver_manager = import_optional("adbc_driver_manager")
-            adbc_version = parse_version(
-                getattr(adbc_driver_manager, "__version__", "0.0")
+            from polars.io.database._utils import (
+                _get_adbc_module_name_from_uri,
+                _import_optional_adbc_driver,
+                _open_adbc_connection,
             )
-            from polars.io.database._utils import _open_adbc_connection
+
+            conn, can_close_conn = (
+                (_open_adbc_connection(connection), True)
+                if isinstance(connection, str)
+                else (connection, False)
+            )
+
+            driver_manager = import_optional("adbc_driver_manager")
+            driver_manager_str_version = getattr(driver_manager, "__version__", "0.0")
+            driver_manager_version = parse_version(driver_manager_str_version)
 
             if if_table_exists == "fail":
                 # if the table exists, 'create' will raise an error,
                 # resulting in behaviour equivalent to 'fail'
                 mode = "create"
             elif if_table_exists == "replace":
-                if adbc_version < (0, 7):
-                    adbc_str_version = ".".join(str(v) for v in adbc_version)
-                    msg = f"`if_table_exists = 'replace'` requires ADBC version >= 0.7, found {adbc_str_version}"
+                if driver_manager_version < (0, 7):
+                    msg = (
+                        "`if_table_exists = 'replace'` requires ADBC version >= 0.7, "
+                        f"found {driver_manager_str_version}"
+                    )
                     raise ModuleUpgradeRequiredError(msg)
                 mode = "replace"
             elif if_table_exists == "append":
@@ -4293,36 +4303,60 @@ class DataFrame:
                 )
                 raise ValueError(msg)
 
-            conn, can_close_conn = (
-                (_open_adbc_connection(connection), True)
-                if isinstance(connection, str)
-                else (connection, False)
-            )
             with (
                 conn if can_close_conn else contextlib.nullcontext(),  # type: ignore[union-attr]
                 conn.cursor() as cursor,  # type: ignore[union-attr]
             ):
                 catalog, db_schema, unpacked_table_name = unpack_table_name(table_name)
                 n_rows: int
-                if adbc_version >= (0, 7):
-                    if "sqlite" in conn.adbc_get_info()["driver_name"].lower():  # type: ignore[union-attr]
-                        if if_table_exists == "replace":
-                            # note: adbc doesn't (yet) support 'replace' for sqlite
-                            cursor.execute(f"DROP TABLE IF EXISTS {table_name}")
-                            mode = "create"
-                        catalog, db_schema = db_schema, None
 
+                adbc_module_name = (
+                    _get_adbc_module_name_from_uri(connection)
+                    if isinstance(connection, str)
+                    else connection_module_root
+                )
+                adbc_driver = _import_optional_adbc_driver(
+                    adbc_module_name, dbapi_submodule=False
+                )
+                adbc_driver_str_version = getattr(adbc_driver, "__version__", "0.0")
+                adbc_driver_version = parse_version(adbc_driver_str_version)
+
+                if adbc_module_name.split("_")[-1] == "sqlite":
+                    catalog, db_schema = db_schema, None
+
+                    # note: ADBC didnt't support 'replace' until adbc-driver-sqlite
+                    # version 0.11 (it was released for other drivers in version 0.7)
+                    if (
+                        driver_manager_version >= (0, 7)
+                        and adbc_driver_version < (0, 11)
+                        and if_table_exists == "replace"
+                    ):
+                        cursor.execute(f"DROP TABLE IF EXISTS {table_name}")
+                        mode = "create"
+
+                # use of schema-qualified table names was released in
+                # adbc-driver-manager 0.7.0 and is working without bugs from driver
+                # version (e.g., adbc-driver-postgresql) version 0.8.0
+                if driver_manager_version >= (0, 7) and adbc_driver_version >= (0, 8):
+                    # As of adbc_driver_manager 1.6.0, adbc_ingest can take a Polars
+                    # DataFrame via the PyCapsule interface
+                    data = self if driver_manager_version >= (1, 6) else self.to_arrow()
                     n_rows = cursor.adbc_ingest(
                         unpacked_table_name,
-                        data=self.to_arrow(),
+                        data=data,
                         mode=mode,
                         catalog_name=catalog,
                         db_schema_name=db_schema,
                         **(engine_options or {}),
                     )
                 elif db_schema is not None:
-                    adbc_str_version = ".".join(str(v) for v in adbc_version)
-                    msg = f"use of schema-qualified table names requires ADBC version >= 0.8, found {adbc_str_version}"
+                    adbc_driver_pypi_name = adbc_module_name.replace("_", "-")
+                    msg = (
+                        "use of schema-qualified table names requires "
+                        "adbc-driver-manager version >= 0.7.0, found "
+                        f"{driver_manager_str_version} and {adbc_driver_pypi_name} "
+                        f"version >= 0.8.0, found {adbc_driver_str_version}"
+                    )
                     raise ModuleUpgradeRequiredError(
                         # https://github.com/apache/arrow-adbc/issues/1000
                         # https://github.com/apache/arrow-adbc/issues/1109
